+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.AndroidSlide"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
