+++ app/src/main/res/layout-sw600dp/activity_main.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="activity"
-            type="com.glacier.androidslide.MainActivity" />
+            type="com.glacier.androidslide.ui.MainActivity" />
 
         <variable
             name="viewModel"
@@ -102,7 +102,7 @@
             app:setSlide="@{slide}"
             app:slide="@{slide}" />
 
-        <com.glacier.androidslide.DrawingView
+        <com.glacier.androidslide.ui.DrawingView
             android:id="@+id/dv_drawing"
             android:layout_width="0dp"
             android:layout_height="0dp"
